server/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
django-allauth==65.3.0
dj-rest-auth==7.0.0
djangorestframework-simplejwt==5.3.1
